study_0254/requirements.txt
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
numpy>=1.19.0
rclpy>=0.10.0
sensor-msgs>=2.2.0
cv-bridge>=3.0.0
image-transport>=2.3.0
tensorflow>=2.8.0
tensorflow-hub>=0.12.0
torch>=1.10.0
transformers>=4.30.0
Pillow>=8.0.0 